requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
chainmap==1.0.3
charset-normalizer==3.4.6
colorama==0.4.6
combomethod==1.0.12
coverage==7.13.5
cryptography==46.0.5
Faker==40.11.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
jdk==0.0.6
nulltype==2.3.1
options==1.4.10
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PySocks==1.7.1
pytest==9.0.2
pytest-cov==7.1.0
python-dotenv==1.2.2
requests==2.32.5
selenium==4.41.0
service-identity==24.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
